+++ src/components/sign/Naver.jsx
@@ -1,4 +1,5 @@
 import React, { useEffect } from "react";
+import { useSelector } from "react-redux";
 
 const Naver = () => {
   const { naver } = window;
@@ -6,15 +7,12 @@ const Naver = () => {
   const naverSignin = () => {
     const naverLogin = new naver.LoginWithNaverId({
       clientId: "r6F2Dw1s6vUkQEJeB2eS",
-      callbackUrl: "http://localhost:3000",
+      callbackUrl: "http://localhost:3000/login",
       isPopup: false,
       loginButton: { color: "white", type: 2, height: "40" },
       callbackHandle: true,
-      accessToken: localStorage.getItem("naverToken"),
     });
     naverLogin.init();
-
-    console.log(naverLogin);
   };
 
   useEffect(() => {
